make migrations correction

+++ database/migrations/2019_07_23_102758_create_type_produitcommande_table.php
@@ -15,7 +15,7 @@ class CreateTypeProduitcommandeTable extends Migration
     {
         Schema::create('type_produitcommande', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->bigIncrements('nbproduit');
+            $table->integer('nbproduit');
             $table->integer('produit_id')->unsigned();
             $table->integer('commande_id')->unsigned();
